Always passing indx to PrimoEngine
Our primo server was failing due to the indx not being passed in, so
now regardless of if args[:start] is passed in, the indx is added to
the Primo URL.

The Primo vcr cassettes were regenerated with the new code as the URLs
changed (i.e. &indx=1 is now included). I did have to modify the smoke
test as a few fields weren't in the newly recorded results so the test
was failing.