Confstrut 1.x, conf.engine_id instead of id
Upgrading to confstrut 1.x, keeps us from using "id" as a config key,
https://github.com/intridea/hashie/issues/290
So we change to "engine_id", which is maybe less confusing anyway,
consistent with how we set key in response.

I think this was considered purely internal API so is not considered
a backwards incompat change, the name of the key being id vs engine_id.
I think/hope.