Get atom XSD validation test working again
For reasons unclear, it stopped working to load the atom xsd into nokogiri. It tried to import a subsidiary XSD from http://www.w3.org/2001/03/xml.xsd -- this started silently failing, which then made load of the schema file itself fail. I think maybe because that started redirecting to an `https` URL, and libxml can't import that for us?

Loading that external XSD on every test run is maybe rude anyway.  Best practices for this kind of stuff are unclear, with the general decline of this kind of semantic XML stuff.

To make it work again, I made a local copy of the xml.xsd, and changed the atom.xsd to reference it, and changed the way we load the xsd into nokogiri (use File.open instead of File.read), so nokogiri can resolve the relative reference. Bingo, passes again.